+++ mobx_codegen/test/data/valid_input_annotation_store_config_has_to_string.dart
@@ -10,7 +10,6 @@ typedef VoidCallback = void Function();
 
 class User = UserBase with _$User;
 
-@MakeStoreConfig(makeToString: true)
 abstract class UserBase with Store {
   UserBase(this.id);