+++ mobx_codegen/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+- Added automatic generation of toString method with @StoreConfig annotation ([@hawkbee1](https://github.com/hawkbee1))
+
 ## 1.0.0 - 1.0.1
 
 - Ready for prime time!